fix: BaseEntity 수정
