--- src/main/java/com/sprint/mission/discodeit/entity/BaseEntity.java
@@ -1,41 +1,0 @@
-package com.sprint.mission.discodeit.entity;
-
-import com.sprint.mission.discodeit.global.util.TimeUtil;
-import lombok.Getter;
-
-import java.io.Serializable;
-import java.time.Instant;
-import java.util.Objects;
-import java.util.UUID;
-
-@Getter
-public abstract class BaseEntity implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-    private final UUID id;
-    private final Instant createdAt;
-    private Instant updatedAt;
-
-    public BaseEntity() {
-        this.id = UUID.randomUUID();
-        this.createdAt = TimeUtil.getCurrentTime();
-        this.updatedAt = createdAt;
-    }
-
-    public void updateUpdatedAt() {
-        this.updatedAt = TimeUtil.getCurrentTime();
-    }
-
-    @Override
-    public boolean equals(Object object) {
-        if (this == object) return true;
-        if (object == null || getClass() != object.getClass()) return false;
-        BaseEntity that = (BaseEntity) object;
-        return Objects.equals(id, that.id) && Objects.equals(createdAt, that.createdAt);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
