+++ src/main/java/com/sprint/mission/discodeit/entity/BinaryContent.java
@@ -1,23 +1,18 @@
 package com.sprint.mission.discodeit.entity;
 
-import com.sprint.mission.discodeit.global.util.TimeUtil;
-import java.io.Serializable;
-import java.time.Instant;
-import java.util.Arrays;
+import com.sprint.mission.discodeit.entity.base.BaseEntity;
 import java.util.UUID;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 
 @Getter
-@EqualsAndHashCode
-public class BinaryContent implements Serializable {
+@EqualsAndHashCode(callSuper = true)
+public class BinaryContent extends BaseEntity {
 
-  private static final long serialVersionUID = 1L;
   private final UUID id;
   private String fileName;
   private String contentType;
   private final byte[] content;
-  private final Instant createdAt;
 
   public static final BinaryContent EMPTY = new BinaryContent(
       "temFileName", "Text", new byte[0]);
@@ -27,19 +22,9 @@ public class BinaryContent implements Serializable {
     this.fileName = fileName;
     this.contentType = contentType;
     this.content = content;
-    this.createdAt = TimeUtil.getCurrentTime();
   }
 
   public static BinaryContent of(String fileName, String contentType, byte[] content) {
     return new BinaryContent(fileName, contentType, content);
-  }
-
-  @Override
-  public String toString() {
-    return "BinaryContent{" +
-        "id=" + id +
-        ", content=" + Arrays.toString(content) +
-        ", createdAt=" + createdAt +
-        '}';
   }
 }
